Add support for scaling the extracted features from style image.
Unfortunately, due to an existing bug in original code, this commit breaks the backwards compatibility in the meaning of produced images. Till now, the features were not scaled at all. The content/style images ratio was calculated, but never applied. Now it calculates the ratio and then further scales it by a value given in '--feature-scale' argument (1 by default).